+++ src/config/mod.rs
@@ -1,3 +1,4 @@
+pub mod cache_setup;
 pub mod cli_args;
 pub mod setup;
 pub mod types;
